[GTWTR-601] Adapt libnss-ato to work with eNB
This commit changes libnss-ato Makefile to use PowerPC's gcc and adds
specific flags for compilation in Linux kernel 2.6.33.
This commit changes default user in .conf.
This commit changes libnss_ato code to use the user specified in .conf
(this was needed because "name" contains the name of the user that does
not exist locally, leading to fail to log in to Linux CLI).